Internalize strings in VM
Now only one instance of any string exists in the virtual machine.
Interning or internalizing strings in the VM allows for less duplication
and faster comparison of strings. Inside the VM, strings may be compared
using the equality operator `==` instead of memcmp() since there exists
only one instance of a unique string. As a result, two (or more) uses of
it must refer to the same memory.