Map instructions to line numbers with RLE format
Previously, each bytecode instruction in a chunk stored a line number in
a corresponding array. Because several instruction tend to correspond to
the same line in a file, this design and implementation proved to be an
inefficient use of space.

Now, bytecode instructions in a chunk map to a line number using a
run-length encoding format (RLE). Line numbers map to offsets, and
offsets map to instructions in the chunk, where this offset is the index
of the bytecode instruction in array `code`. As a result, only the first
instruction of a line must create a new line number. If more
instructions exist on the same line, they're mapped to the same line
number based on their offset, as it must be greater than that of the
first instruction in the same line, but less than that of the first
instruction in the next line.

The format doesn't need to be *that* fast since it's outside the
critical path. The line number of an instruction is required only for
debug or error messages. If there are too many of these, there exists a
more critical problem.